requirements.txt
Django==1.11.3
django-crispy-forms==1.6.1
django-markdown-deux==1.0.5
django-pagedown==0.1.3
gunicorn==19.7.1
markdown2==2.3.4
olefile==0.44
Pillow==4.2.0
psycopg2==2.7.1
pytz==2017.2
whitenoise==3.3.0
